Move throwsWith into test_util
Associated cleanup of dependencies
