+++ pubspec.yaml
@@ -29,7 +29,6 @@ dev_dependencies:
   build_runner: ^0.3.0
   build_test: ^0.4.0+1
   code_builder: '1.0.0-beta+3'
-  matcher: '>=0.12.0 <0.12.1'
   mockito: '>=0.11.0 <2.0.0'
   source_gen: ^0.5.4
   quiver: '>=0.21.4 <0.25.0'
