+++ test/compiler/expression_parser/lexer_test.dart
@@ -3,8 +3,9 @@ library angular2.test.compiler.expression_parser.lexer_test;
 
 import "package:angular2/src/compiler/expression_parser/lexer.dart"
     show Lexer, Token;
-import "package:angular2/src/testing/internal.dart";
 import "package:test/test.dart";
+
+import "../../test_util.dart";
 
 List<Token> lex(String text) {
   return new Lexer().tokenize(text);
