+++ test/compiler/expression_parser/parser_test.dart
@@ -6,9 +6,9 @@ import "package:angular2/src/compiler/expression_parser/ast.dart"
 import "package:angular2/src/compiler/expression_parser/lexer.dart" show Lexer;
 import "package:angular2/src/compiler/expression_parser/parser.dart"
     show Parser;
-import "package:angular2/src/testing/internal.dart";
 import "package:test/test.dart";
 
+import "../../test_util.dart";
 import "unparser.dart" show Unparser;
 
 void main() {
