+++ test/core/application_ref_test.dart
@@ -36,6 +36,7 @@ import 'package:logging/logging.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 
+import "../test_util.dart";
 import 'core_mocks.dart';
 
 void main() {
