+++ test/core/di/map_injector_test.dart
@@ -2,8 +2,9 @@
 library angular2.test.core.di.map_injector_test;
 
 import "package:angular2/core.dart" show Injector, MapInjectorFactory;
-import "package:angular2/src/testing/internal.dart";
 import 'package:test/test.dart';
+
+import "../../test_util.dart";
 
 void main() {
   group("MapInjector", () {
