+++ test/core/linker/view_injector_integration_test.dart
@@ -1,10 +1,12 @@
 @TestOn('browser && !js')
 library angular2.test.core.linker.view_injector_integration_test;
 
-import "package:angular2/core.dart";
 import 'package:angular2/common.dart';
+import "package:angular2/core.dart";
 import "package:angular2/src/testing/internal.dart";
 import 'package:test/test.dart';
+
+import "../../test_util.dart";
 
 const ALL_DIRECTIVES = const [
   SimpleDirective,
