+++ test/core/reflection/reflector_test.dart
@@ -6,9 +6,9 @@ import "package:angular2/src/core/reflection/reflection.dart"
     show Reflector, ReflectionInfo;
 import "package:angular2/src/core/reflection/reflection_capabilities.dart"
     show ReflectionCapabilities;
-import "package:angular2/src/testing/internal.dart";
 import 'package:test/test.dart';
 
+import "../../test_util.dart";
 import "reflector_common.dart"
     show
         ClassDecorator,
